+++ __tests__/lib/mdxish/mdxishMdastToMd.test.ts
@@ -1,7 +1,7 @@
 import type { Root as MdastRoot, RootContent } from 'mdast';
 
 import { NodeTypes } from '../../../enums';
-import { mdxishMdastToMd } from '../../../lib/mdxish';
+import { mdxishMdastToMd } from '../../../lib';
 
 describe('mdxishMdastToMd', () => {
   it('should convert a simple paragraph', () => {
